app/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Eugene Cho <eugenecho@cohere.com>"]
readme = "README.md"
packages = [{include = "backend"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
cohere = "^5.13.11"
google-genai = "^1.2.0"
python-dotenv = "^1.0.1"
matplotlib = "^3.10.0"
sympy = "^1.13.3"
antlr4-python3-runtime = "4.11"
streamlit = "^1.42.1"
flask = "^3.1.0"
flask-cors = "^5.0.1"
werkzeug = "^3.1.3"
pyjwt = "^2.10.1"
redis = "^5.2.1"
psycopg2-binary = "^2.9.10"
celery = "^5.2.7"
waitress = "^3.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
